16/12/2020 casa

+++ Programacion1/Python/ej_bucles/media.py
@@ -3,5 +3,5 @@ numeros. '''
 cantidadn = int(input("Cuantos numeros va a introducir"))
 suma=0
 for i in range(0,cantidadn):
-    suma = suma+int(input("Introduce un numero"))
+    suma = suma+float(input("Introduce un numero"))
 print(suma/cantidadn)
